mid stride massive refactor almost there

+++ scripts/publish_roads.py
@@ -11,9 +11,9 @@ import datetime
 import arcpy
 from arcgis.gis import GIS
 from config import Config
-from scripts.portal import PortalContent, show
+from scripts.portal import PortalContent
 sys.path.insert(0,'')
-from utils import getServiceItem
+
 
 def find_map(aprx, mapname):
     maps = aprx.listMaps(mapname)
@@ -34,7 +34,7 @@ def create_service_definition(map: object, item: dict, sd_file: str):
         service_name=item["pkgname"],
     )
 
-    sddraft.federatedServerUrl = Config.SERVER_URL
+    sddraft.federatedServerUrl = Config.SERVER_AGS # Using an URL here often fails.
     sddraft.description = item["description"]
     sddraft.credits = Config.CREDITS
     sddraft.useLimitations = Config.DISCLAIMER_TEXT
@@ -42,18 +42,17 @@ def create_service_definition(map: object, item: dict, sd_file: str):
     sddraft.portalFolder = item["folder"]
     # sddraft.summary = "say something here"
     # sddraft.serverFolder
-    sddraft.overwriteExistingService = True # WELL THIS FAILS SPECTACULARLY
+    sddraft.overwriteExistingService = True # THIS FAILS EVERY TIME IF THE FILE EXISTS
 
     # The draft file is an XML file.
     sddraft_file = os.path.join(Config.SCRATCH_WORKSPACE, item["pkgname"] + ".sddraft")
 
+    if os.path.exists(sddraft_file): os.unlink(sddraft_file)
+    if os.path.exists(sd_file): os.unlink(sd_file)
+
     sddraft.exportToSDDraft(sddraft_file)
 
-    if os.path.exists(sd_file):
-        #return sd_file # Uncomment this line for faster debugging.
-        os.unlink(sd_file)
-
-    print("Creating", sd_file)
+    print("    Creating", sd_file)
     # "Stage" is the Esri verb for "archive with 7Z".
     # The 7z file contains a FGDB and various settings.
     arcpy.server.StageService(sddraft_file, sd_file)
@@ -77,16 +76,19 @@ if __name__ == "__main__":
 
     (scriptpath, scriptname) = os.path.split(__file__)
 
-    arcpy.env.workspace = Config.SCRATCH_WORKSPACE
-
+    overwrite = True
     initials = os.environ.get("USERNAME")[0:2].upper()
     textmark = (
         datetime.datetime.now().strftime("%m/%d/%y %H:%M") + " " + initials
     )  # more readable
 
-    portal = GIS(Config.PORTAL_URL, Config.PORTAL_USER, Config.PORTAL_PASSWORD)
-    print("%s Logged in as %s" % (textmark, str(portal.properties.user.username)))
-    pc = PortalContent(portal)
+    arcpy.env.workspace = Config.SCRATCH_WORKSPACE
+
+    gis = GIS(Config.PORTAL_URL, Config.PORTAL_USER, Config.PORTAL_PASSWORD)
+    portal = PortalContent(gis)
+    print("%s Logged in as %s" % (textmark, str(portal.gis.properties.user.username)))
+    # Validate the group list.
+    release_groups = portal.getGroups(Config.RELEASE_GROUP_LIST)
 
     try:
         aprx = arcpy.mp.ArcGISProject(Config.BASEMAP_APRX)
@@ -105,7 +107,7 @@ if __name__ == "__main__":
             "description": """<p>This map is used for queries.</p>"""
             + layer_desc
             + project_desc
-            + "<br />" + Config.AUTOGRAPH,
+            + "<br />" + Config.DOC_LINK,
             #"folder": "TESTING_Brian", "pkgname": "DELETEME_" + Config.ROAD_MAP.replace(" ", "_"), # TEST RUN
             "folder": "Public Works", 
             "pkgname": Config.ROAD_MAP.replace(" ", "_"), # SHOWTIME!!!!
@@ -113,7 +115,7 @@ if __name__ == "__main__":
     ]
 
     # Validate the group list
-    staging_groups = pc.get_groups(Config.STAGING_GROUP_LIST)
+    release_groups = portal.getGroups(Config.RELEASE_GROUP_LIST)
 
     total = len(mapnames)
 
@@ -125,16 +127,16 @@ if __name__ == "__main__":
 
         map = find_map(aprx, mapname)
         if not map:
-            print('ERROR! Map not found in APRX. Skipping "%s".' % mapname)
+            print(f'ERROR! Map "{mapname}" not found in APRX.')
             continue
 
         # https://pro.arcgis.com/en/pro-app/2.9/help/sharing/overview/automate-sharing-web-layers.htm
 
         sd_file = os.path.join(Config.SCRATCH_WORKSPACE, item["pkgname"] + ".sd")
-        if arcpy.Exists(sd_file):
+        if arcpy.Exists(sd_file) and not overwrite:
             print(f"Using existing file \"{sd_file}\".")
         else:
-            print(progress, 'Building "%s".' % item["pkgname"])
+            print(progress, f'Building "{item["pkgname"]}" sd file in {Config.SCRATCH_WORKSPACE}.')
             
             # 1. Create a service definition draft
             # 2. Stage the service (zip)
@@ -145,7 +147,7 @@ if __name__ == "__main__":
         #  * First I should try to update an existing definition.
         #  * If that fails then I create a new one...
 
-        print('Uploading definition using "%s" to "%s" folder.' % (item["pkgname"], item["folder"]))
+        print(f'    Uploading sd using "{item["pkgname"]}" to "{item["folder"]}" folder.')
 
         # Upload the service definition to SERVER
         # In theory everything needed to publish the service is already in the SD file.
@@ -154,23 +156,21 @@ if __name__ == "__main__":
         try:
             # in_startupType HAS TO BE "STARTED" else no service is started on the SERVER.
             rval = arcpy.SignInToPortal(Config.PORTAL_URL, Config.PORTAL_USER, Config.PORTAL_PASSWORD)
-            rval = arcpy.server.UploadServiceDefinition(sd_file, Config.SERVER_URL, in_startupType="STARTED")
+            rval = arcpy.server.UploadServiceDefinition(sd_file, Config.SERVER_AGS, in_startupType="STARTED")
 
         except Exception as e:
             print("Could not generate service.", e)
             if e.args[0].startswith("ERROR 001117"):
-                print(
-                    'ERROR. Open the APRX file in ArcGIS Pro and put a description in "%s".'
-                    % mapname
-                )
-                print('Skipping "%s".' % mapname)
+                print(f'ERROR: Open the APRX file in ArcGIS Pro and put a description in properties for "{mapname}".')
+                print(f'Skipping "{mapname}".')
             continue
 
-        # Add a comment to the service
+        # Skipping this step bit me, (2023-02-09) because suddenly everyone had to log in.
+        portal.updateSharing(mapname, everyone=True, groups=release_groups)
 
-        gis = GIS(Config.PORTAL_URL, username=Config.PORTAL_USER, password=Config.PORTAL_PASSWORD)            
-        target_item = getServiceItem(gis, PortalContent(gis), item["pkgname"])
+        # Add a comment to the service
+        target_item = portal.getServiceItem(item["pkgname"])
         target_item.add_comment("Updated %s" % textmark)
-        show(target_item)
+        PortalContent.show(target_item)
 
     print("All done!!!")
